[Refactor] Stats 엔티티를 stats 도메인으로 이동
